@@ -1,9 +1,7 @@
 package com.ycm.work5.Activity;
 
-
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-
+import androidx.appcompat.app.AppCompatActivity;
 import com.ycm.work5.R;
 
 public class MainActivity extends AppCompatActivity {
