@@ -2,21 +2,25 @@ package com.ycm.work5.Activity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.view.View;
 import android.view.animation.Animation;
 import android.view.animation.AnimationUtils;
 import android.widget.ImageView;
-
+import androidx.appcompat.app.AppCompatActivity;
 import com.ycm.work5.R;
 
-public class SplashActivity extends AppCompatActivity {
+import butterknife.BindView;
+import butterknife.ButterKnife;
 
+public class SplashActivity extends AppCompatActivity {
+@BindView(R.id.tv_loading)
+ImageView tv_loading;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_splash);
-        ImageView LoadingView = findViewById(R.id.tv_loading);
+
+        ButterKnife.bind(this);
+
         Animation animation = AnimationUtils.loadAnimation(SplashActivity.this, R.anim.anim);
         animation.setFillAfter(true);
         animation.setAnimationListener(new Animation.AnimationListener() {
@@ -36,6 +40,6 @@ public class SplashActivity extends AppCompatActivity {
 
             }
         });
-        LoadingView.startAnimation(animation);
+        tv_loading.startAnimation(animation);
     }
 }
