@@ -1,24 +1,37 @@
 package com.ycm.work5.Fragment;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-
-/*import androidx.annotation.NonNull;
+import android.widget.TextView;
+import butterknife.ButterKnife;
+import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;*/
-
+import androidx.fragment.app.Fragment;
 import com.ycm.work5.R;
+import butterknife.BindView;
+import butterknife.OnClick;
+import butterknife.Unbinder;
 
 public class CartFragment extends Fragment {
+    @BindView(R.id.tv_gw)
+    TextView tv_gw;
+    Unbinder unbinder;
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.fragment_cart, container, false);
+        unbinder= ButterKnife.bind(this,view);
         return view;
     }
+
+    @Override
+    public void onDestroyView() {
+        super.onDestroyView();
+        unbinder.unbind();
+    }
+  @OnClick(R.id.tv_gw)
+  public void onViewClicked() {
+  }
 }
